refactor: composant tableau mutualisé
